test(node): websocket 存活测试并入 proto_connection_test
archtest 的 `TestGoTestFilesFollowSourceFiles` 要求每个 _test.go 有同名
源文件，而 `websocket_liveness_test.go` 没有 `websocket_liveness.go`,
所以 master 的 test job 一直是红的(与本 PR 无关,顺手带上)。

这三个用例测的 `webSocketEnvelopeStream`、`newWebSocketEnvelopeStream`、
`dialProtoWebSocket`、`websocket*` 常量全部定义在 proto_connection.go,
并入 proto_connection_test.go 就是归位。纯文件移动,用例本身一行未改。
